mesh: Introduced a rigorous primer for accessor expressions.
